requirements.txt
fastapi
uvicorn[standard]
chardet
pycryptodome
pydantic
pydantic_core
pyln-client
jinja2